Afegit control de variable indefinida.
git-svn-id: http://devel.cpl.upc.edu/recursos/svn/smeagol/branches/tiquet_383@1932 4ab63a03-c2d3-4037-8559-9a21b594da05

+++ lib/V2/Server/Controller/Booking.pm
@@ -926,7 +926,7 @@ sub default_DELETE {
     my $req = $c->request;
     
 if ($id) {
-    if(($module eq 'tag') && ($id_module)){
+    if((defined $module) && ($module eq 'tag') && ($id_module)){
         $c->detach( 'delete_relation_tag_booking', [$id, $id_module]);
     }
     else {
